Move manage.py commands into a cli blueprint
